+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -458,6 +458,7 @@ sub build_isf_public_interface_contract {
             't/1289-isf-aggregate-drive-call-values.t',
             't/1290-isf-aggregate-drive-call-expression-values.t',
             't/1291-isf-aggregate-inline-drive-values.t',
+            't/1292-isf-aggregate-inline-drive-expression-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',